WIP: tweaks on border-policy-no-backend
