+++ border-policy-no-backend/src/mlp.rs
@@ -2,7 +2,7 @@ use crate::Mat;
 use serde::{Deserialize, Serialize};
 
 #[cfg(feature = "tch")]
-use tch::nn::VarStore;
+mod tch;
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
 /// Multilayer perceptron with ReLU activation function.
@@ -25,20 +25,5 @@ impl Mlp {
             }
         }
         x.tanh()
-    }
-
-    #[cfg(feature = "tch")]
-    pub fn from_varstore(vs: &VarStore, w_names: &[&str], b_names: &[&str]) -> Self {
-        let vars = vs.variables();
-        let ws: Vec<Mat> = w_names
-            .iter()
-            .map(|name| vars[&name.to_string()].copy().into())
-            .collect();
-        let bs: Vec<Mat> = b_names
-            .iter()
-            .map(|name| vars[&name.to_string()].copy().into())
-            .collect();
-
-        Self { ws, bs }
     }
 }
